pyproject.toml
[tool.poetry]
name = "drf-educ"
version = "0.1.0"
description = ""
authors = ["GLEF1X <glebgar567@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
asgiref = "^3.5.2"
charset-normalizer = "^3.0.0"
Django = { version = "^4.1.1", extras = ["argon2"]}
django-axes = "^5.39.0"
django-behaviors = "^0.5.1"
django-environ = "^0.9.0"
django-ipware = "^4.0.2"
django-split-settings = "^1.1.0"
django-storages = "^1.13.1"
djangorestframework-camel-case = "^1.3.0"
drf-spectacular = {extras = ["sidecar"], version = "^0.24.2"}
drf-spectacular-sidecar = "^2022.11.1"
sentry-sdk = "^1.9.9"
whitenoise = "^6.2.0"
gunicorn = "^20.1.0"
django-health-check = "^3.17.0"
psycopg2-binary = "^2.9.5"
django-baton = "^2.4.1"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.1"}
django-cors-headers = "^3.13.0"
Pillow = "^9.2.0"
python-json-logger = "^2.0.4"
django-structlog = "^4.0.0"
django-extensions = "^3.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
